Run test scripts in meson CI
The autofoo build is on its way out and I am too lazy to properly
integrate these scripts with the meson build. Thus, this commit just
adds them to one single CI build. At least those that can relatively
easily be made to work. The others are left as comments.

Signed-off-by: Uli Schlachter <psychon@znc.in>