CI: Invoke test scripts via `meson test` where possible
Using `meson test` is the Meson equivalent of `make check`, and is the
way developers of other Meson-built projects will generally expect to
run tests. In particular, it encapsulates which of the tests want to be
run from which directories, removing the need for the Gitlab-CI
integration to know this.

Signed-off-by: Simon McVittie <smcv@debian.org>